remove: 팔로우 관련 코드 이동
