+++ src/pages/ProfilePage/index.tsx
@@ -2,12 +2,12 @@ import styled from "styled-components";
 import ProfileView from "./Views/ProfileView";
 import ProfilePostsView from "./Views/ProfilePostsView";
 import { useNavigate, useParams } from "react-router-dom";
-import { useMutation, useQuery } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { GET_USER, ME } from "@/constants/queryKey";
 import { _FOLLOW, _GET_USER, _UNFOLLOW } from "@/api/queries/profile";
 import { useEffect, useState } from "react";
 import { IUser } from "@/types";
-import { IFollow, IUnFollow } from "@/types/profile";
+
 import { useUI } from "@/components/common/uiContext";
 import { _GET, rootAPI } from "@/api";
 
@@ -18,16 +18,8 @@ const ProfileWrap = styled.div`
 const ProfilePage = () => {
   const { id } = useParams();
   const [userData, setUserData] = useState<IUser>();
-  const [isFollow, setIsFollow] = useState(false);
-  const [followId, setFollowId] = useState<string>();
-  const [formData, setFormData] = useState<IFollow | IUnFollow>();
 
-  const { openModal, setModalView, displayModal } = useUI();
-
-  const { data: myData, refetch } = useQuery({
-    queryKey: [ME],
-    queryFn: async () => await _GET("/auth-user"),
-  });
+  const { openModal, setModalView } = useUI();
 
   const navigate = useNavigate();
 
@@ -38,33 +30,6 @@ const ProfilePage = () => {
         "application/x-www-form-urlencoded";
     };
   }, []);
-
-  const followMutation = useMutation({
-    mutationFn: async (formData: IFollow) => {
-      return await _FOLLOW(formData);
-    },
-    onSuccess(data) {
-      setIsFollow(true);
-      setFollowId(data._id);
-    },
-    onError(error) {
-      console.error("ERROR: 팔로우 실패", error);
-    },
-  });
-
-  const unfollowMutation = useMutation({
-    mutationFn: async (formData: IUnFollow) => {
-      return await _UNFOLLOW(formData);
-    },
-    onSuccess() {
-      console.log("API 성공! UnFollow");
-      setIsFollow(false);
-      setFollowId("");
-    },
-    onError(error) {
-      console.error("ERROR: 언팔로우 실패", error);
-    },
-  });
 
   const { data, isLoading, error, isSuccess } = useQuery({
     queryKey: [GET_USER, id],
@@ -85,17 +50,6 @@ const ProfilePage = () => {
     openModal();
   };
 
-  useEffect(() => {
-    if (displayModal) return;
-
-    const refetchData = async () => {
-      const { data: refetchData } = await refetch();
-      setUserData(refetchData?.data);
-    };
-
-    refetchData();
-  }, [displayModal]);
-
   const handleChangeImage = (e: React.MouseEvent) => {
     e.stopPropagation();
 
@@ -111,22 +65,6 @@ const ProfilePage = () => {
   const handleClickChat = (e: React.MouseEvent, id: string) => {
     e.stopPropagation();
     navigate(`/chat/${id}`);
-  };
-
-  const handleClickFollow = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    if (id === undefined) return;
-
-    setFormData({
-      userId: id,
-    });
-    console.log(myData);
-
-    if (!isFollow) {
-      followMutation.mutate(formData as IFollow);
-    } else {
-      unfollowMutation.mutate(formData as IUnFollow);
-    }
   };
 
   useEffect(() => {
@@ -152,13 +90,11 @@ const ProfilePage = () => {
           <>
             <ProfileView
               userInfo={userData}
-              isFollow={isFollow}
               onClickPassword={handleChangePassword}
               onClickEdit={handleChangeName}
               onClickAvatar={handleChangeImage}
               onClickCover={handleChangeCover}
               onClickChat={handleClickChat}
-              onClickFollow={handleClickFollow}
             />
             <ProfilePostsView posts={userData ? userData.posts : []} />
           </>
